make event card component common and with separate card for coming soon events
